display permission on edit profile click
